Completed the CRUD opeartions for all business models.
Modified the duty to have a reference to the project it has been assigned to.

Modified the project model to use DateTime.Now for its due date for new instances by default.

Added migrations for setting the correct behavior on foreign keys when related items are deleted; specifically the relationship between [projects<->duties] and [employees<->teams].